add tests, mv to utils
